Old_type Usage [Deprecated] Google AI Python SDK for the Gemini API With Gemini 2.0, we took the chance to create a single unified SDK for all developers who want to use Google's GenAI models (Gemini, Veo, Imagen, etc). As part of that process, we took all of the feedback from this SDK and what developers like about other SDKs in the ecosystem to create the Google Gen AI SDK.
The full migration guide from the old SDK to new SDK is available in the Gemini API docs.

The Gemini API docs are fully updated to show examples of the new Google Gen AI SDK. We know how disruptive an SDK change can be and don't take this change lightly, but our goal is to create an extremely simple and clear path for developers to build with our models so it felt necessary to make this change.

Thank you for building with Gemini and let us know if you need any help!

Please be advised that this repository is now considered legacy. For the latest features, performance improvements, and active development, we strongly recommend migrating to the official Google Generative AI SDK for Python.

Support Plan for this Repository:

Limited Maintenance: Development is now restricted to critical bug fixes only. No new features will be added.
Purpose: This limited support aims to provide stability for users while they transition to the new SDK.
End-of-Life Date: All support for this repository (including bug fixes) will permanently end on August 31st, 2025.
We encourage all users to begin planning their migration to the Google Generative AI SDK to ensure continued access to the latest capabilities and support.